wip:more tests
